refactor(analysis-types): split complexity DTOs
